+++ src/roboto/ai/agent_session/__init__.py
@@ -4,6 +4,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+from ..core import AnalysisScope
 from .agent_session import AgentSession
 from .client_tool import ClientTool, client_tool
 from .event import (
@@ -59,6 +60,7 @@ __all__ = [
     "AgentToolResultEvent",
     "AgentToolUseContent",
     "AgentToolUseEvent",
+    "AnalysisScope",
     "ClientTool",
     "ClientToolResult",
     "ClientToolResultStatus",